Use MD5HasherWith16k in par2.Encoder
Also change some delegate methods to take an int64 for byteCount.

+++ par2/create.go
@@ -34,7 +34,7 @@ type CreateDelegate interface {
 type DoNothingCreateDelegate struct{}
 
 // OnDataFileLoad implements the CreateDelegate interface.
-func (DoNothingCreateDelegate) OnDataFileLoad(i, n int, path string, byteCount int, err error) {
+func (DoNothingCreateDelegate) OnDataFileLoad(i, n int, path string, byteCount int64, err error) {
 }
 
 // OnIndexFileWrite implements the CreateDelegate interface.
